move `cast` into `ir`
