+++ src/coherence/solve.rs
@@ -5,7 +5,7 @@ use itertools::Itertools;
 use errors::*;
 use ir::*;
 use rust_ir::*;
-use cast::*;
+use ir::cast::*;
 use solve::{SolverChoice, Solution};
 
 struct DisjointSolver {
