+++ src/solve/slg/implementation.rs
@@ -1,4 +1,4 @@
-use crate::cast::{Cast, Caster};
+use crate::ir::cast::{Cast, Caster};
 use crate::chalk_engine::fallible::Fallible;
 use crate::ir::could_match::CouldMatch;
 use crate::ir::*;
